switch to enum
