+++ AGENTS.md
@@ -324,6 +324,10 @@ Keep this list updated when new protocol features are added to kernel.
 - Prefer `==` over `matches!` for simple single-variant enum comparisons. `matches!` is
   for patterns with bindings or guards. For example: `self == Variant` not
   `matches!(self, Variant)`.
+- Prefer `#[repr(C)]` enums for closed FFI choice sets instead of integer aliases and constants.
+  Add `cbindgen:prefix-with-name=true` so generated variants remain unambiguous. Use an integer
+  discriminator only when unknown values are intentionally recoverable, and validate them at the
+  boundary. Invalid enum tags are undefined behavior, so unsafe FFI APIs must require valid tags.
 - Prefer `StructField::nullable` / `StructField::not_null` over
   `StructField::new(name, type, bool)` when nullability is known at compile time.
   Reserve `StructField::new` for cases where nullability is a runtime value.
